Extract canonical review surface and mention request classification slices.
Move GitHub surface upsert helpers and review-details merge logic out of review.ts, and move mention intent routing/classification out of mention.ts with focused unit tests.
