+++ src/handlers/review.ts
@@ -243,6 +243,13 @@ import {
   buildVisibleBudgetProjectionFromEvidence,
 } from "../review-orchestration/review-visible-budget-evidence.ts";
 import {
+  type CanonicalReviewSurface,
+  type CanonicalSurfaceKind,
+  findCanonicalReviewSurface,
+  upsertCanonicalReviewSurface,
+  upsertDegradedReviewDetailsFallbackComment,
+} from "../review-orchestration/review-canonical-surface.ts";
+import {
   toProductionLogBudgetReasoning,
   toProductionLogCandidateFindingSnapshot,
   toProductionLogHunkEmbeddingCounts,
@@ -818,403 +825,6 @@ async function fetchCommitMessages(
 
 
 
-type CanonicalReviewSurface =
-  | { kind: "issue_comment"; commentId: number; body: string }
-  | { kind: "pull_review"; reviewId: number; body: string };
-
-type CanonicalSurfaceKind = CanonicalReviewSurface["kind"];
-
-function getCanonicalReviewSurfaceId(surface: CanonicalReviewSurface): number {
-  return surface.kind === "issue_comment" ? surface.commentId : surface.reviewId;
-}
-
-async function findCanonicalReviewSurface(params: {
-  octokit: Octokit;
-  owner: string;
-  repo: string;
-  prNumber: number;
-  reviewOutputKey: string;
-  surfaceKind: CanonicalSurfaceKind;
-}): Promise<CanonicalReviewSurface | null> {
-  const marker = buildReviewOutputMarker(params.reviewOutputKey);
-
-  if (params.surfaceKind === "issue_comment") {
-    const commentsResponse = await params.octokit.rest.issues.listComments({
-      owner: params.owner,
-      repo: params.repo,
-      issue_number: params.prNumber,
-      per_page: 100,
-      sort: "created",
-      direction: "desc",
-    });
-
-    const issueComment = commentsResponse.data.find((comment) =>
-      typeof comment.id === "number"
-      && typeof comment.body === "string"
-      && comment.body.includes(marker)
-    );
-    const issueCommentBody = typeof issueComment?.body === "string" ? issueComment.body : undefined;
-
-    if (typeof issueComment?.id === "number" && issueCommentBody !== undefined) {
-      return {
-        kind: "issue_comment",
-        commentId: issueComment.id,
-        body: issueCommentBody,
-      };
-    }
-
-    return null;
-  }
-
-  const reviewsResponse = await params.octokit.rest.pulls.listReviews({
-    owner: params.owner,
-    repo: params.repo,
-    pull_number: params.prNumber,
-    per_page: 100,
-  });
-
-  const pullReview = [...reviewsResponse.data].reverse().find((review) =>
-    typeof review.id === "number"
-    && typeof review.body === "string"
-    && review.body.includes(marker)
-  );
-  const pullReviewBody = typeof pullReview?.body === "string" ? pullReview.body : undefined;
-
-  if (typeof pullReview?.id === "number" && pullReviewBody !== undefined) {
-    return {
-      kind: "pull_review",
-      reviewId: pullReview.id,
-      body: pullReviewBody,
-    };
-  }
-
-  return null;
-}
-
-async function updateCanonicalReviewSurface(params: {
-  octokit: Octokit;
-  owner: string;
-  repo: string;
-  prNumber: number;
-  surface: CanonicalReviewSurface;
-  body: string;
-  botHandles: string[];
-}): Promise<CanonicalReviewSurface> {
-  const sanitizedBody = sanitizeOutgoingMentions(params.body, params.botHandles);
-
-  if (params.surface.kind === "issue_comment") {
-    await params.octokit.rest.issues.updateComment({
-      owner: params.owner,
-      repo: params.repo,
-      comment_id: params.surface.commentId,
-      body: sanitizedBody,
-    });
-
-    return {
-      kind: "issue_comment",
-      commentId: params.surface.commentId,
-      body: sanitizedBody,
-    };
-  }
-
-  await params.octokit.request(
-    "PUT /repos/{owner}/{repo}/pulls/{pull_number}/reviews/{review_id}",
-    {
-      owner: params.owner,
-      repo: params.repo,
-      pull_number: params.prNumber,
-      review_id: params.surface.reviewId,
-      body: sanitizedBody,
-    },
-  );
-
-  return {
-    kind: "pull_review",
-    reviewId: params.surface.reviewId,
-    body: sanitizedBody,
-  };
-}
-
-async function createCanonicalReviewSurface(params: {
-  octokit: Octokit;
-  owner: string;
-  repo: string;
-  prNumber: number;
-  reviewOutputKey: string;
-  surfaceKind: CanonicalSurfaceKind;
-  body: string;
-  botHandles: string[];
-  pullReviewEvent?: "APPROVE" | "COMMENT" | "REQUEST_CHANGES";
-}): Promise<CanonicalReviewSurface> {
-  const sanitizedBody = sanitizeOutgoingMentions(params.body, params.botHandles);
-
-  if (params.surfaceKind === "issue_comment") {
-    const response = await params.octokit.rest.issues.createComment({
-      owner: params.owner,
-      repo: params.repo,
-      issue_number: params.prNumber,
-      body: sanitizedBody,
-    });
-
-    if (typeof response.data.id !== "number") {
-      throw new Error("Created canonical issue comment did not return an id");
-    }
-
-    return {
-      kind: "issue_comment",
-      commentId: response.data.id,
-      body: sanitizedBody,
-    };
-  }
-
-  const response = await params.octokit.rest.pulls.createReview({
-    owner: params.owner,
-    repo: params.repo,
-    pull_number: params.prNumber,
-    event: params.pullReviewEvent ?? "COMMENT",
-    body: sanitizedBody,
-  });
-
-  if (typeof response.data.id === "number") {
-    return {
-      kind: "pull_review",
-      reviewId: response.data.id,
-      body: sanitizedBody,
-    };
-  }
-
-  const createdSurface = await findCanonicalReviewSurface({
-    octokit: params.octokit,
-    owner: params.owner,
-    repo: params.repo,
-    prNumber: params.prNumber,
-    reviewOutputKey: params.reviewOutputKey,
-    surfaceKind: "pull_review",
-  });
-
-  if (createdSurface?.kind === "pull_review") {
-    return createdSurface;
-  }
-
-  throw new Error("Created canonical pull review could not be reloaded");
-}
-
-async function upsertCanonicalReviewSurface(params: {
-  octokit: Octokit;
-  owner: string;
-  repo: string;
-  prNumber: number;
-  reviewOutputKey: string;
-  preferredKind: CanonicalSurfaceKind;
-  body?: string;
-  reviewDetailsBlock?: string;
-  summaryBody?: string;
-  canonicalSurface?: CanonicalReviewSurface;
-  requireDegradationDisclosure?: boolean;
-  reviewBoundedness?: ReviewBoundednessContract | null;
-  botHandles: string[];
-  pullReviewEvent?: "APPROVE" | "COMMENT" | "REQUEST_CHANGES";
-  recheckCanPublish?: () => boolean;
-}): Promise<CanonicalReviewSurface | undefined> {
-  let existingSurface = params.canonicalSurface?.kind === params.preferredKind
-    ? params.canonicalSurface
-    : await findCanonicalReviewSurface({
-      octokit: params.octokit,
-      owner: params.owner,
-      repo: params.repo,
-      prNumber: params.prNumber,
-      reviewOutputKey: params.reviewOutputKey,
-      surfaceKind: params.preferredKind,
-    });
-
-  if (!existingSurface && params.reviewDetailsBlock) {
-    const alternateKind: CanonicalSurfaceKind = params.preferredKind === "issue_comment" ? "pull_review" : "issue_comment";
-    existingSurface = await findCanonicalReviewSurface({
-      octokit: params.octokit,
-      owner: params.owner,
-      repo: params.repo,
-      prNumber: params.prNumber,
-      reviewOutputKey: params.reviewOutputKey,
-      surfaceKind: alternateKind,
-    });
-  }
-
-  const body = params.reviewDetailsBlock
-    ? (() => {
-      const summaryBody = params.summaryBody ?? existingSurface?.body;
-      if (!summaryBody) {
-        throw new Error(`Canonical ${params.preferredKind} surface not found for review output marker`);
-      }
-
-      return mergeReviewDetailsIntoSummaryBody({
-        summaryBody,
-        reviewDetailsBlock: params.reviewDetailsBlock,
-        requireDegradationDisclosure: params.requireDegradationDisclosure ?? false,
-        reviewBoundedness: params.reviewBoundedness,
-      });
-    })()
-    : params.body;
-
-  if (!body) {
-    throw new Error("Canonical review surface upsert requires body content");
-  }
-
-  if (params.recheckCanPublish && !params.recheckCanPublish()) {
-    return undefined;
-  }
-
-  if (existingSurface) {
-    return await updateCanonicalReviewSurface({
-      octokit: params.octokit,
-      owner: params.owner,
-      repo: params.repo,
-      prNumber: params.prNumber,
-      surface: existingSurface,
-      body,
-      botHandles: params.botHandles,
-    });
-  }
-
-  return await createCanonicalReviewSurface({
-    octokit: params.octokit,
-    owner: params.owner,
-    repo: params.repo,
-    prNumber: params.prNumber,
-    reviewOutputKey: params.reviewOutputKey,
-    surfaceKind: params.preferredKind,
-    body,
-    botHandles: params.botHandles,
-    pullReviewEvent: params.pullReviewEvent,
-  });
-}
-
-async function upsertDegradedReviewDetailsFallbackComment(params: {
-  octokit: Awaited<ReturnType<GitHubApp["getInstallationOctokit"]>>;
-  owner: string;
-  repo: string;
-  prNumber: number;
-  reviewOutputKey: string;
-  body: string;
-  botHandles: string[];
-  recheckCanPublish?: () => boolean;
-}): Promise<number | undefined> {
-  const { octokit, owner, repo, prNumber, reviewOutputKey, body, botHandles } = params;
-  const marker = buildReviewDetailsMarker(reviewOutputKey);
-  const sanitizedBody = sanitizeOutgoingMentions(body, botHandles);
-
-  const commentsResponse = await octokit.rest.issues.listComments({
-    owner,
-    repo,
-    issue_number: prNumber,
-    per_page: 100,
-    sort: "created",
-    direction: "desc",
-  });
-
-  const existingComment = commentsResponse.data.find((comment) =>
-    typeof comment.body === "string" && comment.body.includes(marker)
-  );
-
-  if (params.recheckCanPublish && !params.recheckCanPublish()) {
-    return undefined;
-  }
-
-  if (existingComment) {
-    await octokit.rest.issues.updateComment({
-      owner,
-      repo,
-      comment_id: existingComment.id,
-      body: sanitizedBody,
-    });
-    return existingComment.id;
-  }
-
-  const response = await octokit.rest.issues.createComment({
-    owner,
-    repo,
-    issue_number: prNumber,
-    body: sanitizedBody,
-  });
-  return response.data.id;
-}
-
-function unwrapKodiaiResponseDetails(summaryBody: string): string {
-  return summaryBody.replace(
-    /\n?<details>\s*\n?<summary>kodiai response<\/summary>\s*\n+([\s\S]*?)\n<\/details>\n?/,
-    (_match, inner: string) => `\n${inner.trim()}\n`,
-  ).trim();
-}
-
-function ensureVisibleApprovalDecision(summaryBody: string): string {
-  if (!summaryBody.includes("Decision: APPROVE")) {
-    return summaryBody;
-  }
-
-  if (summaryBody.trimStart().startsWith("Decision: APPROVE")) {
-    return summaryBody;
-  }
-
-  const leadingWhitespaceLength = summaryBody.length - summaryBody.trimStart().length;
-  const leadingWhitespace = summaryBody.slice(0, leadingWhitespaceLength);
-  const rest = summaryBody
-    .slice(leadingWhitespaceLength)
-    .replace(/(^|\n)Decision: APPROVE\n*/g, "$1")
-    .trimStart();
-  return `${leadingWhitespace}Decision: APPROVE\n\n${rest}`;
-}
-
-function mergeReviewDetailsIntoSummaryBody(params: {
-  summaryBody: string;
-  reviewDetailsBlock: string;
-  requireDegradationDisclosure: boolean;
-  reviewBoundedness?: ReviewBoundednessContract | null;
-}): string {
-  let updatedReviewDetails = params.reviewDetailsBlock;
-  let summaryBody = ensureVisibleApprovalDecision(
-    unwrapKodiaiResponseDetails(
-      ensureReviewBoundednessDisclosureInSummary(
-        params.summaryBody,
-        params.reviewBoundedness,
-      ),
-    ),
-  );
-  if (params.requireDegradationDisclosure) {
-    summaryBody = ensureSearchRateLimitDisclosureInSummary(summaryBody);
-  }
-
-  const bodyCounts = parseSeverityCountsFromBody(summaryBody);
-  const bodyTotal = bodyCounts.critical + bodyCounts.major + bodyCounts.medium + bodyCounts.minor;
-  if (bodyTotal > 0) {
-    updatedReviewDetails = updatedReviewDetails.replace(
-      /- Findings: (\d+) critical, (\d+) major, (\d+) medium, (\d+) minor/,
-      (_match, c, ma, me, mi) => {
-        const total = {
-          critical: parseInt(c) + bodyCounts.critical,
-          major: parseInt(ma) + bodyCounts.major,
-          medium: parseInt(me) + bodyCounts.medium,
-          minor: parseInt(mi) + bodyCounts.minor,
-        };
-        return `- Findings: ${total.critical} critical, ${total.major} major, ${total.medium} medium, ${total.minor} minor (includes ${bodyTotal} from summary observations)`;
-      }
-    );
-  }
-
-  const existingReviewDetailsPattern = /\n?<details>\s*\n?<summary>Review Details<\/summary>[\s\S]*?<\/details>(?:\n?\s*<!--\s*kodiai:review-details:[^>]+-->)?\n?/;
-  if (existingReviewDetailsPattern.test(summaryBody)) {
-    return summaryBody.replace(existingReviewDetailsPattern, `\n\n${updatedReviewDetails}\n\n`).trim();
-  }
-
-  const closingTag = '</details>';
-  const firstCloseIdx = summaryBody.indexOf(closingTag);
-  if (firstCloseIdx === -1) {
-    return `${summaryBody}\n\n${updatedReviewDetails}`;
-  }
-
-  const insertionIdx = firstCloseIdx + closingTag.length;
-  const before = summaryBody.slice(0, insertionIdx).trimEnd();
-  const after = summaryBody.slice(insertionIdx).trimStart();
-  return after ? `${before}\n\n${updatedReviewDetails}\n\n${after}` : `${before}\n\n${updatedReviewDetails}`;
-}
 
 function resolveAuthorTier(params: {
   authorLogin: string;
